requirements.txt
pandas==1.4.2
streamlit==0.84.0
plotly.express==0.4.1